server/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
openai
apify-client
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2